+++ fabric/src/main/java/com/texelsaurus/minecraft/chameleon/config/ForgeApiConfig.java
@@ -5,6 +5,9 @@ import fuzs.forgeconfigapiport.api.config.v2.ForgeConfigRegistry;
 import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.fml.config.ModConfig;
 
+import java.util.List;
+import java.util.function.Predicate;
+
 // Implemented via Forge Config API Port
 
 public class ForgeApiConfig implements ChameleonConfig
@@ -55,6 +58,11 @@ public class ForgeApiConfig implements ChameleonConfig
     @Override
     public <T extends Enum<T>> ConfigEntry<T> defineEnum (String name, T defaultValue) {
         return new ForgeConfigEntryEnum<T>(BUILDER).name(name).defaultValue(defaultValue);
+    }
+
+    @Override
+    public <T> ConfigEntry<List<? extends T>> defineList (String name, List<? extends T> defaultList, Predicate<Object> elementValidator) {
+        return new ForgeConfigEntryList<T>(BUILDER).validator(elementValidator).name(name).defaultValue(defaultList);
     }
 
     @Override
@@ -129,6 +137,27 @@ public class ForgeApiConfig implements ChameleonConfig
         @Override
         protected ForgeConfigSpec.ConfigValue<T> define () {
             return builder.defineEnum(name, defaultValue);
+        }
+    }
+
+    public class ForgeConfigEntryList<T> extends ForgeConfigEntry<List<? extends T>>
+    {
+        Predicate<Object> validator;
+
+        public ForgeConfigEntryList (ForgeConfigSpec.Builder builder) {
+            super(builder);
+            validator = x -> true;
+        }
+
+        public ForgeConfigEntryList<T> validator(Predicate<Object> validtor) {
+            if (validtor != null)
+                this.validator = validtor;
+            return this;
+        }
+
+        @Override
+        protected ForgeConfigSpec.ConfigValue<List<? extends T>> define () {
+            return builder.defineList(name, defaultValue, validator);
         }
     }
 }